middleware: add a module to provide middleware for HTTP and Events.
Created the `middleware` module to provide easy-to-use middleware for HTTP and Events handler.

It creates middleware.HTTP and middleware.Events which add the following middleware in the correct order:
 - Adds a tracking ID to the context, either from the incoming headers or creates a new one.
 - Adds a logger to the context.
 - Request or Events logger (produces a log entry at the end of the handler).
 - OpenTelemetry middleware. Creates a span for the handler and adds the span to the context.

The HTTP/Events tracking ID middleware from go-libs/tracking was moved to this module.

Check the docs and the examples for more details.