[MID-164] Process kafka messages sequentially and commit manually
This ensures that message processing is more reliable. By pushing the message handling off to a goroutine, we immediately "acknowledged" the kafka message and if a redeployment or an app crash happened during processing a message, it would have lost the messages.

Additionally, when there are thousands of messages in the kafka topic, they all would have been read into memory and executed (more or less) simultaneously, leading to a very high memory consumption.